Remove visit scheduler related code and config
Move remaining calls to orchestration client and remove scheduler config
