+++ feature.env
@@ -14,7 +14,6 @@ SYSTEM_CLIENT_SECRET=clientsecret
 ORCHESTRATION_API_URL=http://localhost:9091/orchestration
 PRISONER_SEARCH_API_URL=http://localhost:9091/offenderSearch
 PRISON_API_URL=http://localhost:9091/prison
-VISIT_SCHEDULER_API_URL=http://localhost:9091/visitScheduler
 PRISONER_CONTACT_REGISTRY_API_URL=http://localhost:9091/contactRegistry
 WHEREABOUTS_API_URL=http://localhost:9091/whereabouts
 PRISON_REGISTER_API_URL=http://localhost:9091/prisonRegister
