+++ server/data/orchestrationApiClient.ts
@@ -6,9 +6,12 @@ import {
   PageVisitDto,
   PrisonerProfile,
   ReserveVisitSlotDto,
+  SessionCapacity,
+  SessionSchedule,
   SupportType,
   Visit,
   VisitHistoryDetails,
+  VisitSession,
 } from './orchestrationApiTypes'
 import { VisitSessionData } from '../@types/bapv'
 
@@ -140,6 +143,45 @@ export default class OrchestrationApiClient {
     return this.restClient.get({
       path: '/visit-support',
     })
+  }
+
+  // orchestration-sessions-controller
+
+  async getVisitSessions(offenderNo: string, prisonId: string): Promise<VisitSession[]> {
+    return this.restClient.get({
+      path: '/visit-sessions',
+      query: new URLSearchParams({
+        prisonId,
+        prisonerId: offenderNo,
+        // 'min' and 'max' params omitted, so using API default between 2 and 28 days from now
+      }).toString(),
+    })
+  }
+
+  async getSessionSchedule(prisonId: string, date: string): Promise<SessionSchedule[]> {
+    return this.restClient.get({
+      path: '/visit-sessions/schedule',
+      query: new URLSearchParams({
+        prisonId,
+        date,
+      }).toString(),
+    })
+  }
+
+  async getVisitSessionCapacity(
+    prisonId: string,
+    sessionDate: string,
+    sessionStartTime: string,
+    sessionEndTime: string,
+  ): Promise<SessionCapacity> {
+    try {
+      return await this.restClient.get({
+        path: '/visit-sessions/capacity',
+        query: new URLSearchParams({ prisonId, sessionDate, sessionStartTime, sessionEndTime }).toString(),
+      })
+    } catch (error) {
+      return null
+    }
   }
 
   // prisoner-profile-controller
