--- server/data/visitSchedulerApiClient.ts
@@ -1,49 +1,0 @@
-import { URLSearchParams } from 'url'
-import RestClient from './restClient'
-import { VisitSession, SessionCapacity, SessionSchedule } from './orchestrationApiTypes'
-import config, { ApiConfig } from '../config'
-
-export default class VisitSchedulerApiClient {
-  private restClient: RestClient
-
-  constructor(token: string) {
-    this.restClient = new RestClient('visitSchedulerApiClient', config.apis.visitScheduler as ApiConfig, token)
-  }
-
-  async getVisitSessions(offenderNo: string, prisonId: string): Promise<VisitSession[]> {
-    return this.restClient.get({
-      path: '/visit-sessions',
-      query: new URLSearchParams({
-        prisonId,
-        prisonerId: offenderNo,
-        // 'min' and 'max' params omitted, so using API default between 2 and 28 days from now
-      }).toString(),
-    })
-  }
-
-  async getSessionSchedule(prisonId: string, date: string): Promise<SessionSchedule[]> {
-    return this.restClient.get({
-      path: '/visit-sessions/schedule',
-      query: new URLSearchParams({
-        prisonId,
-        date,
-      }).toString(),
-    })
-  }
-
-  async getVisitSessionCapacity(
-    prisonId: string,
-    sessionDate: string,
-    sessionStartTime: string,
-    sessionEndTime: string,
-  ): Promise<SessionCapacity> {
-    try {
-      return await this.restClient.get({
-        path: '/visit-sessions/capacity',
-        query: new URLSearchParams({ prisonId, sessionDate, sessionStartTime, sessionEndTime }).toString(),
-      })
-    } catch (error) {
-      return null
-    }
-  }
-}
